Refactor slide extraction processing modules
