+++ app/workflows/steps/extract-slides/index.ts
@@ -1,5 +1,7 @@
 export * from "./config";
 export * from "./job-monitoring";
-export * from "./manifest-processing";
+export * from "./manifest-fetcher";
+export * from "./slide-persistence";
+export * from "./slide-processor";
 export * from "./status-updates";
 export * from "./stream-emitters";
